// frontend/src/components/Payroll/Model/SalaryAnnexure.ts
// SalaryAnnexure.ts
export interface SalaryAnnexure {
  annexure_id: number;
  user_id: number;
  name: string;
  designation: string;
  department: string;
  job_location: string;
  basic: number;
  hra: number;
  special_allowance: number;
  profession_tax: number;
  total_deductions: number;
  net_salary: number;
  annexure_date: string;
}